presenters deleted
3 class was deleted: Presenter_edit, Presenter_map and Presenter_list for code structure optimization.
