+++ app/src/main/java/com/next/diarycraft/Adapter/RecyclerViewCardAdapter.java
@@ -197,7 +197,7 @@ public class RecyclerViewCardAdapter extends RecyclerView.Adapter<RecyclerViewCa
                 if(contextType==FRAGMENT_TYPE) {
                     switch (fragmentType) {
                         case FILE_TYPE:
-                            ((FileFragment) fragment).presenter.deleteNotebean(noteBean);
+                            ((FileFragment) fragment).deleteNotebean(noteBean);
                             ((FileFragment) fragment).refreshList();
                             break;
                         case SEARCH_TYPE:
